backend/requirements.txt
colorlog==6.7.0
fastapi==0.95.1
passlib==1.7.4
pika==1.3.2
psycopg2==2.9.6
pydantic==1.10.7
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.13
uvicorn==0.22.0
